update search funcition

+++ controllers/delete_monday_gs.php
@@ -1,5 +1,6 @@
 <?php
 	require_once('../models/data_access_helper.php');
+	require_once('../models/myFunction.php');
 
 	// Create an instance of data access helper
 	$db = new DataAccessHelper();
@@ -7,12 +8,17 @@
 	// Connect to database
 	$db->connect();
 
-	$id = $_GET["id"];
-	$sql = "DELETE FROM monday WHERE Id = $id";
+	if(isset($_GET["id"])){
+		$id = test_input($_GET["id"]);
+		$sql = "DELETE FROM monday WHERE Id = $id";
 
-	$check = $db->executeNonQuery($sql);
-	if($check == true) echo 1;
-	else echo 0;
-	
+		if($db->executeNonQuery($sql)) echo 1;
+		else echo -2;
+	}
+	else echo -2;
+
+	// -2 : Lỗi xử lý
+	// 1 : Không có lỗi
+
 	$db->close();
 ?>
